+++ easycord/plugins/__init__.py
@@ -1,0 +1,4 @@
+"""Optional first-party EasyCord plugins."""
+from .levels import LevelsPlugin
+
+__all__ = ["LevelsPlugin"]